requirements.txt
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.30
botocore==1.31.30
certifi==2016.2.28
cffi==1.15.1
click==8.1.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.18
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.11.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.2
